Revert "Feature/endpoints-na update  (#26)"
This reverts commit 62c2d189f8685e96056b643e415fa60d1e490016.
