--- backend/core/serializers.py
@@ -1,10 +1,0 @@
-from rest_framework.serializers import ModelSerializer
-from .models import Machine
-
-class MachineSerializer(ModelSerializer):
-    class Meta:
-        model = Machine
-        fields = (
-            'serial_number', 'model', 'brand', 'purchase_date', 'created_date', 'state',
-            'branch'
-        )
