+++ backend/cronus/urls.py
@@ -14,19 +14,8 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path, include
-from rest_framework import routers
-from core import views
-from core.task import views as taskviews
-
-router = routers.DefaultRouter()
-router.register(r'task_type', taskviews.TaskTypeViewSet)
-router.register(r'maintenance_period', taskviews.MaintenancePeriodViewSet)
+from django.urls import path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include(router.urls)),
-    path('machine/', views.MachineView.as_view()),
-    path('task/', taskviews.TaskView.as_view()),
-    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))
 ]
